requirements.txt
anthropic>=0.40.0
claude-code-sdk>=0.0.10
python-dotenv>=1.0.0
pyyaml>=6.0.1
rich>=13.7.0
